Format Toolchain.swift file
